+++ ribs/visualize/__init__.py
@@ -27,10 +27,10 @@ from ribs.visualize._cvt_archive_3d_plot import cvt_archive_3d_plot
 from ribs.visualize._cvt_archive_heatmap import cvt_archive_heatmap
 from ribs.visualize._grid_archive_heatmap import grid_archive_heatmap
 from ribs.visualize._parallel_axes_plot import parallel_axes_plot
-from ribs.visualize._qdax_repertoire_vis import (qdax_repertoire_3d_plot,
-                                                 qdax_repertoire_heatmap)
 from ribs.visualize._sliding_boundaries_archive_heatmap import \
     sliding_boundaries_archive_heatmap
+from ribs.visualize._visualize_qdax import (qdax_repertoire_3d_plot,
+                                            qdax_repertoire_heatmap)
 
 __all__ = [
     "cvt_archive_3d_plot",
